Update and rename project-1-template.cpp to project-1-Current.cpp
Current progress on the project.
